Updated: Added memory feature. Converted to floats from integers to allow decimals. Cleaned up Code.

+++ header1.h
@@ -1,9 +1,9 @@
 #pragma once
 #define MAX 11
 //initializes main varaiables and the declares the functions for the program.
-int a;
-int b;
-int add(a, b);
-int sub(a, b);
-int mult(a, b);
+float a;
+float b;
+float add(a, b);
+float sub(a, b);
+float mult(a, b);
 float divi(a, b);
